♻️ 🚚 move skeleton and fix no-post-card display #28
