+++ src/app/(home)/page.tsx
@@ -14,7 +14,7 @@ import dayjs from "dayjs";
 import { CheckSquare2 } from "lucide-react";
 
 import { getFollows } from "@/lib/db/follow";
-import { getAllPosts } from "@/lib/db/post";
+import { getAllPosts, getUserLatestPosts } from "@/lib/db/post";
 
 import Avatar from "./_components/avatar";
 import CraeteButton from "./_components/create-button";
@@ -30,16 +30,22 @@ export default async function Home() {
     redirect("/sign-in");
   }
 
-  const [allPosts, follows] = await Promise.all([
+  const [allPosts, follows, latestPost] = await Promise.all([
     getAllPosts(),
     getFollows(user.id),
+    getUserLatestPosts(user.id),
   ]);
 
+  const isPostedToday =
+    latestPost?.created_at.getDate() === new Date().getDate();
+
   return (
-    <Box my={20}>
-      <NoPostCard>
-        <CraeteButton type={"text"} />
-      </NoPostCard>
+    <>
+      {!isPostedToday && (
+        <NoPostCard>
+          <CraeteButton type={"text"} />
+        </NoPostCard>
+      )}
 
       {allPosts.map((post) => (
         <Card
@@ -132,6 +138,6 @@ export default async function Home() {
           </Flex>
         </Card>
       ))}
-    </Box>
+    </>
   );
 }
